requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pillow>=10.0.0
numpy>=1.26.0
tensorflow>=2.15.0
pydantic>=2.6.0
